requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
boto3==1.20.24
python-dotenv==0.19.0

